Fix a bug for reads of zero size.
It turns out that the U-blox can sometimes send us messages
of a bogus length that are way too long.  In those cases, we
could be asking for 0 byte reads, which ASIO will happily
do and then immediately return 0 bytes.  Instead of this, if
we detect that we would have asked for zero bytes, just throw
away the entire message and do a full read.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>