Don't wait for acknowledge in setTimeMode.
The basic problem is that setTimeMode can be called from within
a GPS object callback (particularly when running as a "base").
Trying to reconfigure the Rate and RTCM while in the callback
leads to a deadlock, as those reconfigure calls try to wait for
data to come in, but new data will never come in since we are
still handling the previous read.  The solution here is to not
wait for acknowledgement and just go on, hoping for the best.
It's not an ideal solution, but it solves the problem for now.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>